Add player movement task tests
