+++ engine/src/test/kotlin/rs/dusk/engine/client/update/task/NPCVisualsTaskTest.kt
@@ -1,4 +1,4 @@
-package rs.dusk.engine.client.update
+package rs.dusk.engine.client.update.task
 
 import io.mockk.*
 import org.junit.jupiter.api.Assertions.assertEquals
@@ -6,10 +6,9 @@ import org.junit.jupiter.api.Test
 import org.koin.dsl.module
 import rs.dusk.core.io.read.BufferReader
 import rs.dusk.core.io.write.BufferWriter
-import rs.dusk.engine.client.update.task.PlayerVisualsTask
 import rs.dusk.engine.model.entity.index.Indexed
-import rs.dusk.engine.model.entity.index.player.Player
-import rs.dusk.engine.model.entity.index.player.Players
+import rs.dusk.engine.model.entity.index.npc.NPC
+import rs.dusk.engine.model.entity.index.npc.NPCs
 import rs.dusk.engine.model.entity.index.update.Visual
 import rs.dusk.engine.model.entity.index.update.VisualEncoder
 import rs.dusk.engine.model.entity.index.update.Visuals
@@ -22,7 +21,7 @@ import rs.dusk.utility.get
  * @author Greg Hibberd <greg@greghibberd.com>
  * @since April 26, 2020
  */
-internal class PlayerVisualsTaskTest : KoinMock() {
+internal class NPCVisualsTaskTest : KoinMock() {
 
     private val encoder: VisualEncoder<Visual> = mockk(relaxed = true)
 
@@ -30,95 +29,53 @@ internal class PlayerVisualsTaskTest : KoinMock() {
         every { encoder.mask } returns 0x8
     }
 
-    private val addMasks = intArrayOf(encoder.mask)
-    private val players: PooledMapList<Player> = mockk(relaxed = true)
+    private val npcs: PooledMapList<NPC> = mockk(relaxed = true)
     private val encoderModule = module {
-        single { spyk(PlayerVisualsTask(players, arrayOf(encoder), addMasks)) }
+        single { spyk(NPCVisualsTask(npcs, arrayOf(encoder))) }
     }
     override val modules = listOf(entityListModule, encoderModule)
 
     @Test
     fun `Run runs all in parallel`() {
         // Given
-        val updateTask: PlayerVisualsTask = get()
-        val player: Player = mockk(relaxed = true)
-        every { players.forEach(any()) } answers {
-            arg<(Indexed) -> Unit>(0).invoke(player)
+        val updateTask: NPCVisualsTask = get()
+        val npc: NPC = mockk(relaxed = true)
+        every { npcs.forEach(any()) } answers {
+            arg<(Indexed) -> Unit>(0).invoke(npc)
         }
         val visuals: Visuals = mockk(relaxed = true)
-        every { player.visuals } returns visuals
+        every { npc.visuals } returns visuals
         // When
         updateTask.run()
         // Then
-        coVerify {
-            updateTask.runAsync(player)
+        verify {
+            updateTask.runAsync(npc)
         }
     }
 
     @Test
     fun `Update skips if un-flagged`() {
         // Given
-        val task: PlayerVisualsTask = get()
-        val players: Players = get()
+        val task: NPCVisualsTask = get()
+        val npcs: NPCs = get()
         val visuals: Visuals = mockk(relaxed = true)
-        val player: Player = mockk(relaxed = true)
-        every { player.visuals } returns visuals
-        players.add(0, player)
+        val npc: NPC = mockk(relaxed = true)
+        every { npc.visuals } returns visuals
+        npcs.add(0, npc)
         // When
         every { visuals.flag } returns 0
-        task.runAsync(player)
+        task.runAsync(npc)
         // Then
         verify { visuals.update = null }
         verify(exactly = 0) {
             task.encodeUpdate(any())
-            task.encodeAddition(any())
-        }
-    }
-
-    @Test
-    fun `Update writes addition if any addMasks changed`() {
-        // Given
-        val updateTask: PlayerVisualsTask = get()
-        val visuals: Visuals = mockk(relaxed = true)
-        val player: Player = mockk(relaxed = true)
-        every { player.visuals } returns visuals
-        every { visuals.flag } returns 1
-        every { visuals.flagged(any()) } returns true
-        every { updateTask.encodeUpdate(visuals) } just Runs
-        every { updateTask.encodeAddition(visuals) } just Runs
-        // When
-        updateTask.runAsync(player)
-        // Then
-        verifyOrder {
-            updateTask.encodeUpdate(visuals)
-            updateTask.encodeAddition(visuals)
-            visuals.flag = 0
-        }
-    }
-
-    @Test
-    fun `Update doesn't rewrite addition`() {
-        // Given
-        val updateTask: PlayerVisualsTask = get()
-        val visuals: Visuals = mockk(relaxed = true)
-        val player: Player = mockk(relaxed = true)
-        every { player.visuals } returns visuals
-        every { visuals.flag } returns 1
-        every { visuals.flagged(any()) } returns false
-        // When
-        updateTask.runAsync(player)
-        // Then
-        verify(exactly = 0) { updateTask.encodeAddition(visuals) }
-        verifyOrder {
-            updateTask.encodeUpdate(visuals)
-            visuals.flag = 0
         }
     }
 
     @Test
     fun `Encode flagged update`() {
         // Given
-        val updateTask: PlayerVisualsTask = get()
+        val updateTask: NPCVisualsTask = get()
         val visuals: Visuals = mockk(relaxed = true)
         val mask = 0x8
         every { visuals.flag } returns mask
@@ -137,7 +94,7 @@ internal class PlayerVisualsTaskTest : KoinMock() {
     @Test
     fun `Encode ignores not flagged update`() {
         // Given
-        val updateTask: PlayerVisualsTask = get()
+        val updateTask: NPCVisualsTask = get()
         val visuals: Visuals = mockk(relaxed = true)
         val mask = 0x8
         every { visuals.flag } returns mask
@@ -156,25 +113,9 @@ internal class PlayerVisualsTaskTest : KoinMock() {
     }
 
     @Test
-    fun `Encode addition`() {
-        // Given
-        val updateTask: PlayerVisualsTask = get()
-        val visuals: Visuals = mockk(relaxed = true)
-        every { visuals.aspects[any()] } returns mockk(relaxed = true)
-        // When
-        updateTask.encodeAddition(visuals)
-        // Then
-        verifyOrder {
-            updateTask.writeFlag(any(), addMasks.sum())
-            encoder.encode(any(), any())
-            visuals.addition = any()
-        }
-    }
-
-    @Test
     fun `Write small flag`() {
         // Given
-        val updateTask: PlayerVisualsTask = get()
+        val updateTask: NPCVisualsTask = get()
         val writer = BufferWriter()
         // When
         updateTask.writeFlag(writer, 0x10)
@@ -186,7 +127,7 @@ internal class PlayerVisualsTaskTest : KoinMock() {
     @Test
     fun `Write medium flag`() {
         // Given
-        val updateTask: PlayerVisualsTask = get()
+        val updateTask: NPCVisualsTask = get()
         val writer = BufferWriter()
         // When
         updateTask.writeFlag(writer, 0x100)
@@ -199,14 +140,14 @@ internal class PlayerVisualsTaskTest : KoinMock() {
     @Test
     fun `Write large flag`() {
         // Given
-        val updateTask: PlayerVisualsTask = get()
+        val updateTask: NPCVisualsTask = get()
         val writer = BufferWriter()
         // When
         updateTask.writeFlag(writer, 0x10000)
         // Then
         val reader = BufferReader(writer.buffer.array())
         assertEquals(0x80, reader.readUnsignedByte())
-        assertEquals(0x8, reader.readUnsignedByte())
+        assertEquals(0x80, reader.readUnsignedByte())
         assertEquals(0x1, reader.readUnsignedByte())
     }
 }
